Minimize onboarding flow to 6 steps instead of 7
