--- src/app/onboarding/form-components/OnboardingSocialsForm.tsx
@@ -1,89 +1,0 @@
-import FormInput from "@/components/ui/FormInput";
-import React from "react";
-import { useForm } from "react-hook-form";
-
-function OnboardingSocialsForm({
-  onBack,
-  onNext,
-  defaultValues, // new optional prop
-}: {
-  onBack: () => void;
-  onNext: (data: OnboardingSocialsFormData) => void;
-  defaultValues?: Partial<OnboardingSocialsFormData>; // optional partial
-}) {
-  const { register, handleSubmit } = useForm<OnboardingSocialsFormData>({
-    defaultValues,
-  });
-
-  const onSubmit = (data: OnboardingSocialsFormData) => {
-    onNext(data);
-  };
-
-  return (
-    <form
-      className="mt-14 flex flex-col gap-y-3"
-      onSubmit={handleSubmit(onSubmit)}
-    >
-      <h2 className="heading-6 mb-3">Socials</h2>
-
-      {/* LinkedIn */}
-      <div className="flex w-full flex-col gap-x-20 gap-y-2">
-        <label htmlFor="linkedin">LinkedIn URL (Optional)</label>
-        <FormInput
-          type="text"
-          placeholder="https://www.linkedin.com/in/your-name"
-          {...register("linkedin")}
-        />
-      </div>
-
-      {/* Twitter */}
-      <div className="flex w-full flex-col justify-between gap-x-20 gap-y-2">
-        <label htmlFor="twitter">Twitter URL (Optional)</label>
-        <FormInput
-          type="text"
-          placeholder="https://twitter.com/yourhandle"
-          {...register("twitter")}
-        />
-      </div>
-
-      {/* GitHub / GitLab */}
-      <div className="flex w-full flex-col justify-between gap-x-20 gap-y-2">
-        <label htmlFor="git">GitHub/GitLab URL (Optional)</label>
-        <FormInput
-          type="text"
-          placeholder="https://github.com/yourusername"
-          {...register("git")}
-        />
-      </div>
-
-      {/* Personal Website */}
-      <div className="flex w-full flex-col justify-between gap-x-20 gap-y-2">
-        <label htmlFor="personalWebsite">Personal Website URL (Optional)</label>
-        <FormInput
-          type="text"
-          placeholder="https://yourportfolio.com"
-          {...register("personalWebsite")}
-        />
-      </div>
-
-      {/* Navigation Buttons */}
-      <div className="mt-6 flex justify-between">
-        <button
-          type="button"
-          onClick={onBack}
-          className="cursor-pointer rounded border border-white px-4 py-2 text-white"
-        >
-          Back
-        </button>
-        <button
-          type="submit"
-          className="cursor-pointer rounded bg-white px-4 py-2 text-black"
-        >
-          Next
-        </button>
-      </div>
-    </form>
-  );
-}
-
-export default OnboardingSocialsForm;
